Student: Button zum Freigeben hinzugefuegt; kommentiert
Dozent: ist nun ein Controller; kommentiert

Allgemein: Javadoc angelegt, kommentiert
