+++ Hausarbeit-KbSE/src/main/webapp/login.xhtml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE html>
+<!-- Wiho Wert-->
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:h="http://xmlns.jcp.org/jsf/html"
       xmlns:p="http://primefaces.org/ui">
@@ -7,15 +8,18 @@
     <h:head>   
         <title> Osca-MP-Anmeldung </title>        
     </h:head>  
-
+    
+    <!-- CSS fuer Icons -->
     <h:outputStylesheet name="primeicons/primeicons.css" library="primefaces"/>
 
     <h:body style="background-image: url('data/blurred.jpg'); background-repeat: no-repeat; background-size: cover">
 
+        <!-- Logo der Hochschule -->
         <div style="text-align: center">
             <img height="30%" width="30%" src="data/Logo.png" alt="Hier sollte eigentlich ein Bild sein ..." onclick="location.href = 'load.xhtml'"></img>
         </div>
 
+        <!-- Rahmen -->
         <div style="background-color:#FFFFFF; margin:30px; text-align: center; border-left: 6px solid #BCCBD0; border-top: 6px solid #BCCBD0" >
 
             <h:form style="text-align: left;padding: 45px;" id="formID">
@@ -26,6 +30,7 @@
 
                 <p:row>
 
+                    <!-- Grid zum Anmelden eines Studenten oder Dozenten (automatische Weiterleitung an jeweilige Seite) -->
                     <h:panelGrid columns="2" style="margin-bottom:10px" cellpadding="5">
 
                         <p:outputLabel for="username" value="Benutzername: " style="font-family: Andale Mono, monospace; font-size: 14px"/>
