+++ Hausarbeit-KbSE/src/main/webapp/modulAuswahlStudent.xhtml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE html>
+<!-- Wiho Wert-->
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:h="http://xmlns.jcp.org/jsf/html"
       xmlns:p="http://primefaces.org/ui"
@@ -8,31 +9,37 @@
     <h:head>
         <title> Osca-MP-Anmeldung </title>        
     </h:head>  
-    
+
+    <!-- CSS fuer Icons -->
     <h:outputStylesheet name="primeicons/primeicons.css" library="primefaces"/>
-    
+
     <h:body style="background-image: url('data/blurred.jpg'); background-repeat: no-repeat; background-size: cover">
 
+        <!-- Logo der Hochschule -->
         <div style="text-align: center">
             <img height="30%" width="30%" src="data/Logo.png" alt="Hier sollte eigentlich ein Bild sein ..." onclick="location.href = 'load.xhtml'"></img>
         </div>
 
+        <!-- Rahmen -->
         <div style="background-color:#FFFFFF; margin:30px; text-align: center; border-left: 6px solid #BCCBD0; border-top: 6px solid #BCCBD0" >
-            
+
             <h:form style="text-align: left;padding: 45px;" id="form">
-                
+
+                <!-- Anzeigen einer Nachricht -->
                 <p:growl id="msgs" showDetail="true" />
-               
+
+                <!-- Titel --> 
                 <p:row>
                     <h3 style="font-family: Andale Mono, monospace; margin-bottom: 50px">Auswahl des Moduls zur muendlichen Pruefung</h3>
                 </p:row>
 
                 <p:row>
-                    
+
+                    <!-- Grid fuer das Anmelden zu einer Pruefung -->
                     <h:panelGrid columns="2" style="margin-bottom:10px" cellpadding="5">
-                        
+
                         <p:outputLabel for="moduls" value="Module: " style="font-family: Andale Mono, monospace; font-size: 14px"/>
-                        
+
                         <p:selectOneMenu id="moduls" value="#{dropdownViewStudent.modul}" style="width:150px; font-family: Andale Mono, monospace; font-size: 14px">
                             <p:ajax update="moduls"/>
                             <f:selectItem itemLabel="Modul auswaehlen" itemValue="" noSelectionOption="true" />
@@ -40,7 +47,7 @@
                         </p:selectOneMenu>
 
                         <p:outputLabel for="day" value="Tag: " style="font-family: Andale Mono, monospace; font-size: 14px"/>
-                        
+
                         <p:selectOneMenu id="day" value="#{dropdownViewStudent.day}" style="width:150px; font-family: Andale Mono, monospace; font-size: 14px">
                             <f:selectItem itemLabel="Zeit auswaehlen" itemValue="" noSelectionOption="true" />
                             <f:selectItems value="#{dropdownViewStudent.days}" />
@@ -48,19 +55,26 @@
                         </p:selectOneMenu>
 
                         <p:outputLabel for="time" value="Zeit: " style="font-family: Andale Mono, monospace; font-size: 14px"/>
-                        
+
                         <p:selectOneMenu id="time" value="#{dropdownViewStudent.time}" style="width:150px; font-family: Andale Mono, monospace; font-size: 14px">
                             <f:selectItem itemLabel="Zeit auswaehlen" itemValue="" noSelectionOption="true" />
                             <f:selectItems value="#{dropdownViewStudent.times}" />
                             <p:ajax update="time"/>
                         </p:selectOneMenu>
 
-                        <p:commandButton value="Bestaetigen" update="msgs" action="#{dropdownViewStudent.displayLog}" 
-                                         style="font-family: Andale Mono, monospace; font-size: 15px; margin-top: 100px"
-                                         icon="pi pi-check-circle">
-                        </p:commandButton>
                         <p:spacer />
 
+                    </h:panelGrid>
+                    
+                    <h:panelGrid columns="2" style="margin-bottom:10px" cellpadding="5">
+                        <p:commandButton value="Bestaetigen" update="msgs" action="#{dropdownViewStudent.displayLog}" 
+                                         style="font-family: Andale Mono, monospace; font-size: 15px;"
+                                         icon="pi pi-check-circle">
+                        </p:commandButton>
+                        <p:commandButton value="Termin freigeben" update="msgs" action="#{dropdownViewStudent.displayLog}" 
+                                         style="font-family: Andale Mono, monospace; font-size: 15px; color: red"
+                                         icon="pi pi-times-circle">
+                        </p:commandButton>
                     </h:panelGrid>
                 </p:row>
             </h:form>
